WIP: moving code from mathics.builtin.drawing to mathics.eval.drawing
